Merged PR 76594: Add support for computing a completion list
This PR adds a new IWorkspaceServer.GetCompletionList(...) API that returns a list of completion items given a source file and a position within the source file.

Please forgive me for how much code clean up and organization there is in this PR. Much of that was part of the process of learning the code base.