Merged PR 1146439: Try to fix the 503 error in instrumentation
The problem was that the instrumentation syntax tree was not using the same language as present in the compile options. Due to this sometimes the language versions are different with the workspace and the syntax tree which throws the exception.

Also, this PR adds "langVersion" flag to in the .trydotnet file for the command line arguments. This will enable us to be consistent on what language version is being used,

However there is a restore.sh file that also sets the language version, so everytime we update the language versions we must take care to do it for both the files.