Merged PR 1152269: Do not throw an exception when instrumentation is enabled and there are compile errors in the users code
Before trying to instrument, we must check if there are compile errors in the code. If yes then we simply return the compilation and the relevant error messages will hence be displayed to the user.

Also added a test to verify the same and some refactoring around how we check if there was a diagnostic with error severity(both for diagnostic and SerializableDiagnostic)