Merged PR 1279088: WIP: Publish an npm package that includes approval json files
1. Adds an npm package that contains:
a. All the .approved.json files
b. A function that knows how to obtain the contents of these files given a name of the file

This package will be installed in the orchestrator and used in the simulator to verify if the simulator and the agent contracts are in sync

The corresponding PR for the orchestrator: https://msazure.visualstudio.com/DefaultCollection/One/_git/MLS-Agent/pullrequest/1279088?_a=overview

To do:
- [ ] Automate the creation of npm package
- [ ] When will the npm package be created, everytime we do a merge?