Merged PR 1337640: Remove the file system dependency from the CodeBlockParser
Added an InMemoryDirectoryAccessor so that in our tests we just rely upon the inMemoryDirectory thing and not access the file system.

It moves the CodeLinkBlockParser to use the IDirectoryAccessor, hence the corresponding tests have been writeen against the inMemory structure. However we still need to move the markdown project to use thsis interface